# PERF: regression in DataFrame reduction ops performance #37081