@ukarroum 2020-10-14
> do we have an asv that covers this case, if not can you add one?
> 
> this doesn't need a note as its on master.

I believe we do : 

https://pandas.pydata.org/speed/pandas/#stat_ops.FrameOps.time_op?p-op='sum'&p-dtype='int'