@jorisvandenbossche 2020-10-14
Indeed. Jeff, see the issue, it was actually catched thanks to our asv suite