@ukarroum 2020-10-17
Should i do something about the : 2 failed azure pipelines ?
Looks like they're failing on master too and the `./ci/code_checks.sh` localy return no error.